+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- `Innmind\Server\Control\Server\Command::overSsh()`
+
 ### Changed
 
 - Requires PHP `8.4`